+++ packages/common/src/editors/inputEditor.ts
@@ -147,6 +147,9 @@ export class InputEditor implements Editor {
   }
 
   focus(): void {
+    // always set focus on grid first so that plugin to copy range (SlickCellExternalCopyManager) would still be able to paste at that position
+    this.grid.focus();
+
     if (this._input) {
       this._input.focus();
     }