+++ examples/webpack-demo-vanilla-bundle/CHANGELOG.md
@@ -4,6 +4,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.1](https://github.com/ghiscoding/slickgrid-universal/compare/v2.6.0...v2.6.1) (2023-02-24)
+
+**Note:** Version bump only for package webpack-demo-vanilla-bundle
+
+# [2.6.0](https://github.com/ghiscoding/slickgrid-universal/compare/v2.5.0...v2.6.0) (2023-02-23)
+
+**Note:** Version bump only for package webpack-demo-vanilla-bundle
+
 # [2.5.0](https://github.com/ghiscoding/slickgrid-universal/compare/v2.4.1...v2.5.0) (2023-02-17)
 
 ### Bug Fixes